Changed maven source directory, abd added project target files
