--- libs/jar/setup.txt
@@ -1,11 +1,0 @@
-Getting started
-
-Dependencies:
-
-Unirest (Installation procedures here: https://github.com/Mashape/unirest-java#installing)
-PaystackJava installation:
-
-Download jar file here:
-Add jar file as a Module to your Java project:
-On Intellij IDEA: File -> Project Structure -> Modules -> Dependencies Tab -> Add -> JARs or Directories -> Attach jar
-On Netbeans: Project properties -> Libraries -> Compile -> ADD JAR/folder -> Add Jar
